+++ minime/admin.py
@@ -1,3 +1,17 @@
 from django.contrib import admin
+from minime import models
 
-# Register your models here.
+
+class UrlAdmin(admin.ModelAdmin):
+    list_display = [
+        "hash",
+        "created",
+        "has_password",
+    ]
+
+    @admin.display(boolean=True, description="Has a password?")
+    def has_password(self, obj) -> bool:
+        return obj.password is not None
+
+
+admin.register(models.Url, UrlAdmin)